+++ src/utils/mod.rs
@@ -1,1 +1,4 @@
 pub mod input;
+pub mod get_info;
+pub mod generate;
+pub mod loops;
